Install Blender plugin

+++ Editor/Scripts/DCCIntegration/BlenderIntegrator.cs
@@ -1,6 +1,8 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using Unity.AnimeToolbox;
 using Unity.SharpZipLib.Utils;
+using UnityEngine;
 
 namespace UnityEditor.MeshSync {
 internal class BlenderIntegrator : BaseDCCIntegrator {
@@ -14,30 +16,101 @@ internal class BlenderIntegrator : BaseDCCIntegrator {
 //----------------------------------------------------------------------------------------------------------------------
     protected override bool  ConfigureDCCToolV(DCCToolInfo dccToolInfo, string localPluginPath) 
     {
-        //[TODO-sin: 2020-5-7] Implement this
-        //Execute the following
-        //Might be different for different versions of Blender
-        // import bpy
-        // bpy.ops.preferences.addon_install(overwrite=True, filepath="/Users/shin/G/UT/MeshSyncDCCPlugin/Plugins~/Dist/UnityMeshSync_0.1.0-preview_Blender_Mac/blender-2.80.zip")
-        // bpy.ops.preferences.addon_enable(module='unity_mesh_sync')
-        // bpy.ops.wm.save_userpref()
-        
-        
-
         string tempPath = FileUtil.GetUniqueTempPathInProject();
         
         Directory.CreateDirectory(tempPath);
         ZipUtility.UncompressFromZip(localPluginPath, null, tempPath);
         
-        //Cleanup
-        FileUtility.DeleteFilesAndFolders(tempPath);
+        //Go down one folder
+        string extractedPath = null;
+        foreach (string dir in Directory.EnumerateDirectories(tempPath)) {
+            extractedPath = dir;
+            break;
+        }
+
+        if (string.IsNullOrEmpty(extractedPath))
+            return false;
+
+        string ver = dccToolInfo.DCCToolVersion;
+        string pluginFile = Path.Combine(extractedPath, $"blender-{ver}.zip");
         
-        return false;
+        if (!File.Exists(pluginFile)) {
+            return false;
+        }
+        
+        //script
+        string installScriptFileName = $"InstallBlenderPlugin_{ver}.py";
+        string templatePath = Path.Combine(MeshSyncEditorConstants.DCC_INSTALL_SCRIPTS_PATH,installScriptFileName );
+        if (!File.Exists(templatePath)) {
+            return false;
+        }
+
+       
+        //Replace the path in the template with actual path
+        string installScriptFormat = File.ReadAllText(templatePath);
+        string installScript = String.Format(installScriptFormat,pluginFile);
+        string installScriptPath = Path.Combine(tempPath, installScriptFileName);
+        File.WriteAllText(installScriptPath, installScript);
+      
+        bool setupSuccessful = SetupAutoLoadPlugin(dccToolInfo.AppPath, installScriptPath);
+
+        //Cleanup
+        File.Delete(installScriptPath);
+        FileUtility.DeleteFilesAndFolders(tempPath);
+
+        return setupSuccessful;
     }
 
 //----------------------------------------------------------------------------------------------------------------------
     protected override void FinalizeDCCConfigurationV() {
-        //[TODO-sin: 2020-5-12] Implement this
+        DCCToolInfo dccToolInfo = GetDCCToolInfo();
+        
+        EditorUtility.DisplayDialog("MeshSync",
+            $"MeshSync plugin installed for {dccToolInfo.GetDescription()}", 
+            "Ok"
+        );
+    }
+
+//----------------------------------------------------------------------------------------------------------------------
+
+    
+    bool SetupAutoLoadPlugin(string appPath, string installScriptPath) {
+
+        try {
+            if (!System.IO.File.Exists(appPath)) {
+                Debug.LogError("[MeshSync] No Blender installation found at " + appPath);
+                return false;
+            }
+
+            System.Diagnostics.Process process = new System.Diagnostics.Process {
+                StartInfo = {
+                    FileName = appPath,
+                    // WindowStyle = System.Diagnostics.ProcessWindowStyle.Hidden,
+                    // CreateNoWindow = true,
+                    UseShellExecute = false,
+                    RedirectStandardError = true,
+                    Arguments = $"-b -P {installScriptPath}"         //Execute batch mode
+                },
+                EnableRaisingEvents = true
+            };
+            process.Start();
+
+            string stderr = process.StandardError.ReadToEnd();
+            process.WaitForExit();
+            int exitCode = process.ExitCode;
+            
+            if(exitCode != 0 && !string.IsNullOrEmpty(stderr)) {
+                Debug.LogError($"Installation error. ExitCode: {exitCode}: {stderr}");
+                return false;
+            }
+            
+            
+        } catch (Exception e) {
+            Debug.LogError("[MeshSync] Failed to install plugin. Exception: " + e.Message);
+            return false;
+        }
+
+        return true;
     }
